fix Follow Model: rewite Follow model in notepad/models.py to accounts/models.py
