+++ app/config/urls.py
@@ -22,7 +22,6 @@ from django.conf.urls.static import static
 
 
 urlpatterns = [
-    # path('admin/', admin.site.urls),
     path('', include('notepad.urls')),
     path('', include('accounts.urls')),
 ]
