+++ app/notepad/tests/test_views.py
@@ -392,30 +392,6 @@ class QuestionReviewViewTests(TestCase):
         # register review
 
 
-class FollowViewTests(TestCase):
-    """FollowViewのテスト"""
-
-    # 正常値
-    def test_normal_value_view(self):
-        """FollowView: 正常値"""
-        # create user
-        create_user(self)
-        follow_user = User.objects.create_user('follow_test', password='follow_test')
-        self.client.login(username='test_user', password='password')
-        # get request
-        kwargs = {
-            'following': self.user.pk,
-            'followed': follow_user.pk,
-        }
-        url = reverse('notepad:follow', kwargs=kwargs)
-        assert_normal_get_request(self, url, status_code=302)
-        # check redirect
-        redirect_url = reverse('notepad:dashboard', kwargs={'pk': follow_user.pk})
-        resp = self.client.get(url)
-        self.assertRedirects(resp, redirect_url)
-        # check follow
-
-
 class StarViewTests(TestCase):
     """StarViewのテスト"""
